cleanup
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@1548 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/emulationsettings/EmulationSettings.java
@@ -424,10 +424,10 @@ public class EmulationSettings extends C64Window {
 				for (int fc = 0; fc < SIDChip.FC_MAX; fc += STEP) {
 					if (filter.isReSIDFilter6581() || filter.isReSIDFilter8580()) {
 						double data = resid_builder.resid.FilterModelConfig.estimateFrequency(filter, fc);
-						series.getData().add(new XYChart.Data<Number, Number>(fc, (int) data));
+						series.getData().add(new XYChart.Data<Number, Number>(fc, data));
 					} else if (filter.isReSIDfpFilter6581() || filter.isReSIDfpFilter8580()) {
 						double data = resid_builder.residfp.FilterModelConfig.estimateFrequency(filter, fc);
-						series.getData().add(new XYChart.Data<Number, Number>(fc, (int) data));
+						series.getData().add(new XYChart.Data<Number, Number>(fc, data));
 					}
 				}
 			}
